+++ src/main/java/com/woowacourse/pelotonbackend/mission/presentation/dto/MissionUpdateRequest.java
@@ -1,0 +1,4 @@
+package com.woowacourse.pelotonbackend.mission.presentation.dto;
+
+public class MissionUpdateRequest {
+}
